libs/python/pyproject.toml
[project]
name = "eai-common"
version = "0.1.0"
description = "Shared helpers (auth, db, settings) for Enterprise AI platform services"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115",
  "pydantic>=2.6",
  "PyJWT>=2.8",
  "SQLAlchemy>=2.0",
  "psycopg[binary]>=3.1",
]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["eai_common*"]
